+++ python/mlc_llm/protocol/openai_api_protocol.py
@@ -4,13 +4,16 @@ https://github.com/lm-sys/FastChat/blob/main/fastchat/protocol/openai_api_protoc
 """
 
 # pylint: disable=missing-class-docstring
+
+import json
 import time
 from typing import Any, Dict, List, Literal, Optional, Tuple, Union
 
 import shortuuid
 from pydantic import BaseModel, Field, field_validator, model_validator
 
-from mlc_llm.serve.config import ResponseFormat
+from .conversation_protocol import Conversation
+from .error_protocol import BadRequestError
 
 ################ Commons ################
 
@@ -82,7 +85,7 @@ class CompletionRequest(BaseModel):
     """
 
     model: str
-    prompt: Union[str, List[int], List[Union[str, List[int]]]]
+    prompt: Union[str, List[int]]
     best_of: int = 1
     echo: bool = False
     frequency_penalty: float = 0.0
@@ -100,7 +103,7 @@ class CompletionRequest(BaseModel):
     top_p: float = 1.0
     user: Optional[str] = None
     ignore_eos: bool = False
-    response_format: RequestResponseFormat = Field(default_factory=RequestResponseFormat)
+    response_format: Optional[RequestResponseFormat] = None
 
     @field_validator("frequency_penalty", "presence_penalty")
     @classmethod
@@ -214,7 +217,7 @@ class ChatCompletionRequest(BaseModel):
     tool_choice: Optional[Union[Literal["none", "auto"], Dict]] = None
     user: Optional[str] = None
     ignore_eos: bool = False
-    response_format: RequestResponseFormat = Field(default_factory=RequestResponseFormat)
+    response_format: Optional[RequestResponseFormat] = None
 
     @field_validator("frequency_penalty", "presence_penalty")
     @classmethod
@@ -248,6 +251,74 @@ class ChatCompletionRequest(BaseModel):
         if not self.logprobs and self.top_logprobs > 0:
             raise ValueError('"logprobs" must be True to support "top_logprobs"')
         return self
+
+    def check_message_validity(self) -> None:
+        """Check if the given chat messages are valid. Return error message if invalid."""
+        for i, message in enumerate(self.messages):
+            if message.role == "system" and i != 0:
+                raise BadRequestError(
+                    f"System prompt at position {i} in the message list is invalid."
+                )
+            if message.role == "tool":
+                raise BadRequestError("Tool as the message author is not supported yet.")
+            if message.tool_call_id is not None:
+                if message.role != "tool":
+                    raise BadRequestError("Non-tool message having `tool_call_id` is invalid.")
+            if isinstance(message.content, list):
+                if message.role != "user":
+                    raise BadRequestError("Non-user message having a list of content is invalid.")
+            if message.tool_calls is not None:
+                if message.role != "assistant":
+                    raise BadRequestError("Non-assistant message having `tool_calls` is invalid.")
+                raise BadRequestError("Assistant message having `tool_calls` is not supported yet.")
+
+    def check_function_call_usage(self, conv_template: Conversation) -> None:
+        """Check if function calling is used and update the conversation template.
+        Return error message if invalid request format for function calling.
+        """
+
+        # return if no tools are provided or tool_choice is set to none
+        if self.tools is None or (isinstance(self.tool_choice, str) and self.tool_choice == "none"):
+            conv_template.use_function_calling = False
+            return
+
+        # select the tool based on the tool_choice if specified
+        if isinstance(self.tool_choice, dict):
+            if self.tool_choice["type"] != "function":  # pylint: disable=unsubscriptable-object
+                raise BadRequestError("Only 'function' tool choice is supported")
+
+            if len(self.tool_choice["function"]) > 1:  # pylint: disable=unsubscriptable-object
+                raise BadRequestError("Only one tool is supported when tool_choice is specified")
+
+            for tool in self.tools:  # pylint: disable=not-an-iterable
+                if (
+                    tool.function.name
+                    == self.tool_choice["function"][  # pylint: disable=unsubscriptable-object
+                        "name"
+                    ]
+                ):
+                    conv_template.use_function_calling = True
+                    conv_template.function_string = tool.function.model_dump_json()
+                    return
+
+            # pylint: disable=unsubscriptable-object
+            raise BadRequestError(
+                f"The tool_choice function {self.tool_choice['function']['name']}"
+                " is not found in the tools list"
+            )
+            # pylint: enable=unsubscriptable-object
+
+        if isinstance(self.tool_choice, str) and self.tool_choice != "auto":
+            raise BadRequestError(f"Invalid tool_choice value: {self.tool_choice}")
+
+        function_list = []
+        for tool in self.tools:  # pylint: disable=not-an-iterable
+            if tool.type != "function":
+                raise BadRequestError("Only 'function' tool type is supported")
+            function_list.append(tool.function.model_dump())
+
+        conv_template.use_function_calling = True
+        conv_template.function_string = json.dumps(function_list)
 
 
 class ChatCompletionResponseChoice(BaseModel):
@@ -291,6 +362,9 @@ class ChatCompletionStreamResponse(BaseModel):
     model: str
     system_fingerprint: str
     object: Literal["chat.completion.chunk"] = "chat.completion.chunk"
+    usage: UsageInfo = Field(
+        default_factory=lambda: UsageInfo()  # pylint: disable=unnecessary-lambda
+    )
 
 
 ################################################
@@ -315,6 +389,8 @@ def openai_api_get_generation_config(
     request: Union[CompletionRequest, ChatCompletionRequest]
 ) -> Dict[str, Any]:
     """Create the generation config from the given request."""
+    from ..serve.config import ResponseFormat  # pylint: disable=import-outside-toplevel
+
     kwargs: Dict[str, Any] = {}
     arg_names = [
         "n",
@@ -337,5 +413,8 @@ def openai_api_get_generation_config(
         kwargs["max_tokens"] = -1
     if request.stop is not None:
         kwargs["stop_strs"] = [request.stop] if isinstance(request.stop, str) else request.stop
-    kwargs["response_format"] = ResponseFormat(**request.response_format.model_dump(by_alias=True))
+    if request.response_format is not None:
+        kwargs["response_format"] = ResponseFormat(
+            **request.response_format.model_dump(by_alias=True)
+        )
     return kwargs
