+++ python/mlc_llm/serve/__init__.py
@@ -2,10 +2,9 @@
 
 # Load MLC LLM library by importing base
 from .. import base
-from .async_engine import AsyncThreadedEngine
 from .config import EngineMode, GenerationConfig, KVCacheConfig
 from .data import Data, ImageData, RequestStreamOutput, TextData, TokenData
-from .engine import Engine
+from .engine import AsyncEngine, Engine
 from .grammar import BNFGrammar, GrammarStateMatcher
 from .json_schema_converter import json_schema_to_ebnf
 from .request import Request
