+++ python/mlc_llm/serve/data.py
@@ -1,8 +1,9 @@
 """Classes denoting multi-modality data used in MLC LLM serving"""
 
 from dataclasses import dataclass
-from typing import List, Optional, Tuple
+from typing import Dict, List, Optional, Tuple
 
+import tvm
 import tvm._ffi
 from tvm.runtime import Object
 from tvm.runtime.ndarray import NDArray
@@ -80,6 +81,46 @@ class ImageData(Data):
 
     def __len__(self):
         return self.embed_size
+
+    @staticmethod
+    def from_url(url: str, config: Dict) -> "ImageData":
+        """Get the image from the given URL, process and return the image tensor as TVM NDArray."""
+
+        # pylint: disable=import-outside-toplevel, import-error
+        import base64
+        from io import BytesIO
+
+        import requests
+        from PIL import Image
+        from transformers import CLIPImageProcessor
+
+        if url.startswith("data:image"):
+            # The image is encoded in base64 format
+            base64_image = url.split(",")[1]
+            image_data = base64.b64decode(base64_image)
+            image_tensor = Image.open(BytesIO(image_data)).convert("RGB")
+        elif url.startswith("http"):
+            response = requests.get(url, timeout=5)
+            image_tensor = Image.open(BytesIO(response.content)).convert("RGB")
+        else:
+            raise ValueError(f"Unsupported image URL format: {url}")
+
+        image_input_size = config["model_config"]["vision_config"]["image_size"]
+        image_embed_size = (
+            image_input_size // config["model_config"]["vision_config"]["patch_size"]
+        ) ** 2
+
+        image_processor = CLIPImageProcessor(
+            size={"shortest_edge": image_input_size},
+            crop_size={"height": image_input_size, "width": image_input_size},
+        )
+        image_features = tvm.nd.array(
+            image_processor.preprocess(image_tensor, return_tensors="np")["pixel_values"].astype(
+                "float16"
+            )
+        )
+        image_data = ImageData(image_features, image_embed_size)
+        return image_data
 
 
 @dataclass
