+++ python/mlc_llm/serve/entrypoints/openai_entrypoints.py
@@ -2,23 +2,17 @@
 
 # pylint: disable=too-many-locals,too-many-return-statements,too-many-statements
 import ast
-import json
 from http import HTTPStatus
-from typing import AsyncGenerator, Dict, List, Optional, Sequence, Union
+from typing import AsyncGenerator, Dict, List, Optional, Union
 
 import fastapi
 
-from mlc_llm.serve import data
-
-from ...protocol import protocol_utils
-from ...protocol.conversation_protocol import Conversation
-from ...protocol.openai_api_protocol import (
+from mlc_llm.protocol import error_protocol
+from mlc_llm.protocol.openai_api_protocol import (
     ChatCompletionMessage,
     ChatCompletionRequest,
     ChatCompletionResponse,
     ChatCompletionResponseChoice,
-    ChatCompletionStreamResponse,
-    ChatCompletionStreamResponseChoice,
     ChatFunctionCall,
     ChatToolCall,
     CompletionRequest,
@@ -30,8 +24,8 @@ from ...protocol.openai_api_protocol import (
     ModelResponse,
     UsageInfo,
 )
-from ..server import ServerContext
-from . import entrypoint_utils
+from mlc_llm.serve import engine_utils
+from mlc_llm.serve.server import ServerContext
 
 app = fastapi.APIRouter()
 
@@ -59,130 +53,30 @@ async def request_completion(request: CompletionRequest, raw_request: fastapi.Re
     server_context: ServerContext = ServerContext.current()
     async_engine = server_context.get_engine(request.model)
     if async_engine is None:
-        return entrypoint_utils.create_error_response(
+        return error_protocol.create_error_response(
             HTTPStatus.BAD_REQUEST, message=f'The requested model "{request.model}" is not served.'
         )
-    request_id = f"cmpl-{entrypoint_utils.random_uuid()}"
-    async_engine.state.record_event(request_id, event="receive request")
-
-    # - Check if unsupported arguments are specified.
-    error = entrypoint_utils.check_unsupported_fields(request)
-    if error is not None:
-        return error
-
-    # - Process prompt and check validity.
-    async_engine.state.record_event(request_id, event="start tokenization")
-    prompts = entrypoint_utils.process_prompts(request.prompt, async_engine.tokenizer.encode)
-    async_engine.state.record_event(request_id, event="finish tokenization")
-    if isinstance(prompts, fastapi.responses.JSONResponse):
-        # Errored when processing the prompts
-        return prompts
-    if len(prompts) > 1:
-        return entrypoint_utils.create_error_response(
-            HTTPStatus.BAD_REQUEST,
-            message="Entrypoint /v1/completions only accept single prompt. "
-            f"However, {len(prompts)} prompts {prompts} are received.",
-        )
-    error = entrypoint_utils.check_prompts_length(prompts, async_engine.max_input_sequence_length)
-    if error is not None:
-        return error
-    prompt = prompts[0]
-
-    # Process generation config. Create request id.
-    generation_cfg = protocol_utils.get_generation_config(request)
+    request_id = f"cmpl-{engine_utils.random_uuid()}"
 
     # Streaming response.
     if request.stream:
+        # We manually get the first response from generator to
+        # capture potential exceptions in this scope, rather then
+        # the StreamingResponse scope.
+        stream_generator = async_engine._handle_completion(  # pylint: disable=protected-access
+            request, request_id
+        )
+        first_response = await anext(  # type: ignore  # pylint: disable=undefined-variable
+            stream_generator
+        )
 
         async def completion_stream_generator() -> AsyncGenerator[str, None]:
-            # - Echo back the prompt.
-            if request.echo:
-                text = async_engine.tokenizer.decode(prompt)
-                response = CompletionResponse(
-                    id=request_id,
-                    choices=[
-                        CompletionResponseChoice(index=i, text=text)
-                        for i in range(generation_cfg.n)
-                    ],
-                    model=request.model,
-                    usage=UsageInfo(
-                        prompt_tokens=len(prompt),
-                        completion_tokens=0,
-                    ),
-                )
+            if isinstance(first_response, StopAsyncIteration):
+                yield "data: [DONE]\n\n"
+                return
+            yield f"data: {first_response.model_dump_json()}\n\n"
+            async for response in stream_generator:
                 yield f"data: {response.model_dump_json()}\n\n"
-
-            # - Generate new tokens.
-            num_completion_tokens = 0
-            finish_reasons: List[Optional[str]] = [None for _ in range(generation_cfg.n)]
-            async_engine.state.record_event(request_id, event="invoke generate")
-            async for delta_outputs in async_engine.generate(prompt, generation_cfg, request_id):
-                assert len(delta_outputs) == generation_cfg.n
-                choices = []
-                for i, delta_output in enumerate(delta_outputs):
-                    finish_reason_updated = False
-                    if delta_output.finish_reason is not None and finish_reasons[i] is None:
-                        finish_reasons[i] = delta_output.finish_reason
-                        finish_reason_updated = True
-                    num_completion_tokens += delta_output.num_delta_tokens
-                    if not finish_reason_updated and delta_output.delta_text == "":
-                        # Ignore empty delta text when finish reason is not updated.
-                        continue
-
-                    choices.append(
-                        CompletionResponseChoice(
-                            index=i,
-                            finish_reason=finish_reasons[i],
-                            text=delta_output.delta_text,
-                            logprobs=(
-                                LogProbs(
-                                    content=[
-                                        LogProbsContent.model_validate_json(logprob_json_str)
-                                        for logprob_json_str in delta_output.delta_logprob_json_strs
-                                    ]
-                                )
-                                if delta_output.delta_logprob_json_strs is not None
-                                else None
-                            ),
-                        )
-                    )
-
-                if len(choices) == 0:
-                    # Skip yield when there is no delta output.
-                    continue
-                response = CompletionResponse(
-                    id=request_id,
-                    choices=choices,
-                    model=request.model,
-                    usage=UsageInfo(
-                        prompt_tokens=len(prompt),
-                        completion_tokens=num_completion_tokens,
-                    ),
-                )
-                yield f"data: {response.model_dump_json()}\n\n"
-            async_engine.state.record_event(request_id, event="finish")
-
-            # - Echo the suffix.
-            if request.suffix is not None:
-                assert all(finish_reason is not None for finish_reason in finish_reasons)
-                response = CompletionResponse(
-                    id=request_id,
-                    choices=[
-                        CompletionResponseChoice(
-                            index=i,
-                            finish_reason=finish_reason,
-                            text=request.suffix,
-                        )
-                        for i, finish_reason in enumerate(finish_reasons)
-                    ],
-                    model=request.model,
-                    usage=UsageInfo(
-                        prompt_tokens=len(prompt),
-                        completion_tokens=num_completion_tokens,
-                    ),
-                )
-                yield f"data: {response.model_dump_json()}\n\n"
-
             yield "data: [DONE]\n\n"
 
         return fastapi.responses.StreamingResponse(
@@ -190,138 +84,56 @@ async def request_completion(request: CompletionRequest, raw_request: fastapi.Re
         )
 
     # Normal response.
-    init_output_text = "" if not request.echo else async_engine.tokenizer.decode(prompt)
-    output_texts = [init_output_text for _ in range(generation_cfg.n)]
+    num_prompt_tokens = 0
     num_completion_tokens = 0
-    finish_reasons: List[Optional[str]] = [None for _ in range(generation_cfg.n)]
-    logprob_json_strs_list: Optional[List[List[str]]] = (
-        [[] for _ in range(generation_cfg.n)] if generation_cfg.logprobs else None
+    output_texts = ["" for _ in range(request.n)]
+    finish_reasons: List[Optional[str]] = [None for _ in range(request.n)]
+    logprob_results: Optional[List[List[LogProbsContent]]] = (
+        [[] for _ in range(request.n)] if request.logprobs else None
     )
-    async_engine.state.record_event(request_id, event="invoke generate")
-    async for delta_outputs in async_engine.generate(prompt, generation_cfg, request_id):
+
+    async for response in async_engine._handle_completion(  # pylint: disable=protected-access
+        request, request_id
+    ):
         if await raw_request.is_disconnected():
             # In non-streaming cases, the engine will not be notified
             # when the request is disconnected.
             # Therefore, we check if it is disconnected each time,
             # and abort the request from engine if so.
             await async_engine.abort(request_id)
-            return entrypoint_utils.create_error_response(
+            return error_protocol.create_error_response(
                 HTTPStatus.BAD_REQUEST, message="The request has disconnected"
             )
+        num_prompt_tokens = response.usage.prompt_tokens
+        num_completion_tokens = response.usage.completion_tokens
+        for choice in response.choices:
+            output_texts[choice.index] += choice.text
+            if choice.finish_reason is not None and finish_reasons[choice.index] is None:
+                finish_reasons[choice.index] = choice.finish_reason
+            if choice.logprobs is not None:
+                assert logprob_results is not None
+                logprob_results[choice.index] += choice.logprobs.content
 
-        assert len(delta_outputs) == generation_cfg.n
-        for i, delta_output in enumerate(delta_outputs):
-            if delta_output.finish_reason is not None and finish_reasons[i] is None:
-                finish_reasons[i] = delta_output.finish_reason
-            output_texts[i] += delta_output.delta_text
-            num_completion_tokens += delta_output.num_delta_tokens
-            if logprob_json_strs_list is not None:
-                assert delta_output.delta_logprob_json_strs is not None
-                logprob_json_strs_list[i] += delta_output.delta_logprob_json_strs
     assert all(finish_reason is not None for finish_reason in finish_reasons)
-    suffix = request.suffix if request.suffix is not None else ""
-    async_engine.state.record_event(request_id, event="finish")
-    response = CompletionResponse(
+    return CompletionResponse(
         id=request_id,
         choices=[
             CompletionResponseChoice(
                 index=i,
                 finish_reason=finish_reason,
-                text=output_text + suffix,
+                text=output_text,
                 logprobs=(
-                    LogProbs(
-                        content=[
-                            LogProbsContent.model_validate_json(logprob_json_str)
-                            for logprob_json_str in logprob_json_strs_list[  # pylint: disable=unsubscriptable-object
-                                i
-                            ]
-                        ]
-                    )
-                    if logprob_json_strs_list is not None
-                    else None
+                    LogProbs(content=logprob_results[i]) if logprob_results is not None else None
                 ),
             )
             for i, (output_text, finish_reason) in enumerate(zip(output_texts, finish_reasons))
         ],
         model=request.model,
-        usage=UsageInfo(
-            prompt_tokens=len(prompt),
-            completion_tokens=num_completion_tokens,
-        ),
+        usage=UsageInfo(prompt_tokens=num_prompt_tokens, completion_tokens=num_completion_tokens),
     )
-    return response
 
 
 ################ v1/chat/completions ################
-
-
-def chat_completion_check_message_validity(
-    messages: List[ChatCompletionMessage],
-) -> Optional[str]:
-    """Check if the given chat messages are valid. Return error message if invalid."""
-    for i, message in enumerate(messages):
-        if message.role == "system" and i != 0:
-            return f"System prompt at position {i} in the message list is invalid."
-        if message.role == "tool":
-            return "Tool as the message author is not supported yet."
-        if message.tool_call_id is not None:
-            if message.role != "tool":
-                return "Non-tool message having `tool_call_id` is invalid."
-        if isinstance(message.content, list):
-            if message.role != "user":
-                return "Non-user message having a list of content is invalid."
-        if message.tool_calls is not None:
-            if message.role != "assistant":
-                return "Non-assistant message having `tool_calls` is invalid."
-            return "Assistant message having `tool_calls` is not supported yet."
-    return None
-
-
-def check_function_call_usage(
-    request: ChatCompletionRequest, conv_template: Conversation
-) -> Optional[str]:
-    """Check if function calling is used and update the conversation template.
-    Return error message if invalid request format for function calling.
-    """
-
-    # return if no tools are provided or tool_choice is set to none
-    if request.tools is None or (
-        isinstance(request.tool_choice, str) and request.tool_choice == "none"
-    ):
-        conv_template.use_function_calling = False
-        return None
-
-    # select the tool based on the tool_choice if specified
-    if isinstance(request.tool_choice, dict):
-        if request.tool_choice["type"] != "function":
-            return "Only 'function' tool choice is supported"
-
-        if len(request.tool_choice["function"]) > 1:
-            return "Only one tool is supported when tool_choice is specified"
-
-        for tool in request.tools:
-            if tool.function.name == request.tool_choice["function"]["name"]:
-                conv_template.use_function_calling = True
-                conv_template.function_string = tool.function.model_dump_json()
-                return None
-
-        return (
-            f"The tool_choice function {request.tool_choice['function']['name']}"
-            " is not found in the tools list"
-        )
-
-    if isinstance(request.tool_choice, str) and request.tool_choice != "auto":
-        return f"Invalid tool_choice value: {request.tool_choice}"
-
-    function_list = []
-    for tool in request.tools:
-        if tool.type != "function":
-            return "Only 'function' tool type is supported"
-        function_list.append(tool.function.model_dump())
-
-    conv_template.use_function_calling = True
-    conv_template.function_string = json.dumps(function_list)
-    return None
 
 
 def convert_function_str_to_json(stringified_calls: str) -> List[Union[Dict, None]]:
@@ -360,132 +172,30 @@ async def request_chat_completion(
     server_context: ServerContext = ServerContext.current()
     async_engine = server_context.get_engine(request.model)
     if async_engine is None:
-        return entrypoint_utils.create_error_response(
+        return error_protocol.create_error_response(
             HTTPStatus.BAD_REQUEST, message=f'The requested model "{request.model}" is not served.'
         )
-    request_id = f"chatcmpl-{entrypoint_utils.random_uuid()}"
-    async_engine.state.record_event(request_id, event="receive request")
-
-    # - Check if the model supports chat conversation.
-    conv_template = server_context.get_conv_template(request.model)
-    if conv_template is None:
-        return entrypoint_utils.create_error_response(
-            HTTPStatus.BAD_REQUEST,
-            message=f'The requested model "{request.model}" does not support chat.',
-        )
-
-    # - Check if unsupported arguments are specified.
-    error = entrypoint_utils.check_unsupported_fields(request)
-    if error is not None:
-        return error
-
-    # - Process messages and update the conversation template in three steps:
-    #   i. Check the message validity.
-    #  ii. Add the input messages to the conversation template.
-    # iii. Add the additional message for the assistant.
-    error_msg = chat_completion_check_message_validity(request.messages)
-    if error_msg is not None:
-        return entrypoint_utils.create_error_response(HTTPStatus.BAD_REQUEST, message=error_msg)
-
-    # Check for function calling usage and update the conversation template
-    error_msg = check_function_call_usage(request, conv_template)
-    if error_msg is not None:
-        return entrypoint_utils.create_error_response(HTTPStatus.BAD_REQUEST, message=error_msg)
-
-    for message in request.messages:
-        role = message.role
-        content = message.content
-        if role == "system":
-            assert isinstance(content, str)
-            conv_template.system_message = content if content is not None else ""
-            continue
-
-        assert role != "tool", "Internal error: tool role."
-        conv_template.messages.append((role, content))
-    conv_template.messages.append(("assistant", None))
-
-    # - Get the prompt from template, and encode to token ids.
-    # - Check prompt length
-    async_engine.state.record_event(request_id, event="start tokenization")
-
-    model_config = server_context.get_model_config(request.model)
-    prompts = entrypoint_utils.process_prompts(
-        conv_template.as_prompt(model_config),
-        async_engine.tokenizer.encode,
-    )
-
-    async_engine.state.record_event(request_id, event="finish tokenization")
-
-    if conv_template.system_prefix_token_ids is not None:
-        prompts[0] = conv_template.system_prefix_token_ids + prompts[0]
-    error = entrypoint_utils.check_prompts_length(prompts, async_engine.max_input_sequence_length)
-    if error is not None:
-        return error
-
-    prompt: Sequence[Union[List[int], data.ImageData]] = prompts
-
-    # Process generation config. Create request id.
-    generation_cfg = protocol_utils.get_generation_config(
-        request,
-        extra_stop_token_ids=conv_template.stop_token_ids,
-        extra_stop_str=conv_template.stop_str,
-    )
+    request_id = f"chatcmpl-{engine_utils.random_uuid()}"
 
     # Streaming response.
     if request.stream:
+        # We manually get the first response from generator to
+        # capture potential exceptions in this scope, rather then
+        # the StreamingResponse scope.
+        stream_generator = async_engine._handle_chat_completion(  # pylint: disable=protected-access
+            request, request_id
+        )
+        first_response = await anext(  # type: ignore  # pylint: disable=undefined-variable
+            stream_generator
+        )
 
         async def completion_stream_generator() -> AsyncGenerator[str, None]:
-            async_engine.state.record_event(request_id, event="invoke generate")
-            finish_reasons: List[Optional[str]] = [None for _ in range(generation_cfg.n)]
-            async for delta_outputs in async_engine.generate(prompt, generation_cfg, request_id):
-                assert len(delta_outputs) == generation_cfg.n
-                choices = []
-                for i, delta_output in enumerate(delta_outputs):
-                    finish_reason_updated = False
-                    if delta_output.finish_reason is not None and finish_reasons[i] is None:
-                        finish_reasons[i] = (
-                            delta_output.finish_reason
-                            if not conv_template.use_function_calling
-                            else "tool_calls"
-                        )
-                        finish_reason_updated = True
-                    if not finish_reason_updated and delta_output.delta_text == "":
-                        # Ignore empty delta text when finish reason is not updated.
-                        async_engine.state.record_event(request_id, event="skip empty delta text")
-                        continue
-
-                    choices.append(
-                        ChatCompletionStreamResponseChoice(
-                            index=i,
-                            finish_reason=finish_reasons[i],
-                            delta=ChatCompletionMessage(
-                                content=delta_output.delta_text, role="assistant"
-                            ),
-                            logprobs=(
-                                LogProbs(
-                                    content=[
-                                        LogProbsContent.model_validate_json(logprob_json_str)
-                                        for logprob_json_str in delta_output.delta_logprob_json_strs
-                                    ]
-                                )
-                                if delta_output.delta_logprob_json_strs is not None
-                                else None
-                            ),
-                        )
-                    )
-
-                if len(choices) == 0:
-                    # Skip yield when there is no delta output.
-                    continue
-                response = ChatCompletionStreamResponse(
-                    id=request_id,
-                    choices=choices,
-                    model=request.model,
-                    system_fingerprint="",
-                )
-                async_engine.state.record_event(request_id, event="yield delta output")
+            if isinstance(first_response, StopAsyncIteration):
+                yield "data: [DONE]\n\n"
+                return
+            yield f"data: {first_response.model_dump_json()}\n\n"
+            async for response in stream_generator:
                 yield f"data: {response.model_dump_json()}\n\n"
-            async_engine.state.record_event(request_id, event="finish")
             yield "data: [DONE]\n\n"
 
         return fastapi.responses.StreamingResponse(
@@ -493,39 +203,42 @@ async def request_chat_completion(
         )
 
     # Normal response.
-    output_texts = ["" for _ in range(generation_cfg.n)]
+    num_prompt_tokens = 0
     num_completion_tokens = 0
-    finish_reasons: List[Optional[str]] = [None for _ in range(generation_cfg.n)]
-    logprob_json_strs_list: Optional[List[List[str]]] = (
-        [[] for _ in range(generation_cfg.n)] if generation_cfg.logprobs else None
+    output_texts = ["" for _ in range(request.n)]
+    finish_reasons: List[Optional[str]] = [None for _ in range(request.n)]
+    logprob_results: Optional[List[List[LogProbsContent]]] = (
+        [[] for _ in range(request.n)] if request.logprobs else None
     )
-    async_engine.state.record_event(request_id, event="invoke generate")
-    async for delta_outputs in async_engine.generate(prompt, generation_cfg, request_id):
+
+    async for response in async_engine._handle_chat_completion(  # pylint: disable=protected-access
+        request, request_id
+    ):
         if await raw_request.is_disconnected():
             # In non-streaming cases, the engine will not be notified
             # when the request is disconnected.
             # Therefore, we check if it is disconnected each time,
             # and abort the request from engine if so.
             await async_engine.abort(request_id)
-            return entrypoint_utils.create_error_response(
+            return error_protocol.create_error_response(
                 HTTPStatus.BAD_REQUEST, message="The request has disconnected"
             )
+        num_prompt_tokens = response.usage.prompt_tokens
+        num_completion_tokens = response.usage.completion_tokens
+        for choice in response.choices:
+            assert isinstance(choice.delta.content, str)
+            output_texts[choice.index] += choice.delta.content
+            if choice.finish_reason is not None and finish_reasons[choice.index] is None:
+                finish_reasons[choice.index] = choice.finish_reason
+            if choice.logprobs is not None:
+                assert logprob_results is not None
+                logprob_results[choice.index] += choice.logprobs.content
 
-        assert len(delta_outputs) == generation_cfg.n
-        for i, delta_output in enumerate(delta_outputs):
-            if delta_output.finish_reason is not None and finish_reasons[i] is None:
-                finish_reasons[i] = delta_output.finish_reason
-            output_texts[i] += delta_output.delta_text
-            num_completion_tokens += delta_output.num_delta_tokens
-            if logprob_json_strs_list is not None:
-                assert delta_output.delta_logprob_json_strs is not None
-                logprob_json_strs_list[i] += delta_output.delta_logprob_json_strs
     assert all(finish_reason is not None for finish_reason in finish_reasons)
 
-    async_engine.state.record_event(request_id, event="finish")
-
-    tool_calls_list: List[List[ChatToolCall]] = [[] for _ in range(generation_cfg.n)]
-    if conv_template.use_function_calling:
+    tool_calls_list: List[List[ChatToolCall]] = [[] for _ in range(request.n)]
+    use_function_calling = any(finish_reason == "tool_calls" for finish_reason in finish_reasons)
+    if use_function_calling:
         for i, output_text in enumerate(output_texts):
             try:
                 fn_json_list = convert_function_str_to_json(output_text)
@@ -557,20 +270,11 @@ async def request_chat_completion(
                 finish_reason=finish_reasons[i],
                 message=(
                     ChatCompletionMessage(role="assistant", content=output_text)
-                    if (not conv_template.use_function_calling or finish_reason == "error")
+                    if not use_function_calling or finish_reason == "error"
                     else ChatCompletionMessage(role="assistant", tool_calls=tool_calls)
                 ),
                 logprobs=(
-                    LogProbs(
-                        content=[
-                            LogProbsContent.model_validate_json(logprob_json_str)
-                            for logprob_json_str in logprob_json_strs_list[  # pylint: disable=unsubscriptable-object
-                                i
-                            ]
-                        ]
-                    )
-                    if logprob_json_strs_list is not None
-                    else None
+                    LogProbs(content=logprob_results[i]) if logprob_results is not None else None
                 ),
             )
             for i, (output_text, finish_reason, tool_calls) in enumerate(
@@ -579,7 +283,5 @@ async def request_chat_completion(
         ],
         model=request.model,
         system_fingerprint="",
-        usage=UsageInfo(
-            prompt_tokens=sum(len(item) for item in prompt), completion_tokens=num_completion_tokens
-        ),
+        usage=UsageInfo(prompt_tokens=num_prompt_tokens, completion_tokens=num_completion_tokens),
     )
