+++ tests/python/serve/test_serve_async_engine.py
@@ -3,8 +3,8 @@
 import asyncio
 from typing import List
 
-from mlc_llm.serve import AsyncThreadedEngine, GenerationConfig, KVCacheConfig
-from mlc_llm.serve.engine import ModelInfo
+from mlc_llm.serve import AsyncEngine, GenerationConfig, KVCacheConfig
+from mlc_llm.serve.engine_base import ModelInfo
 
 prompts = [
     "What is the meaning of life?",
@@ -28,25 +28,25 @@ async def test_engine_generate():
     )
     kv_cache_config = KVCacheConfig(page_size=16, max_total_sequence_length=4096)
     # Create engine
-    async_engine = AsyncThreadedEngine(model, kv_cache_config)
+    async_engine = AsyncEngine(model, kv_cache_config)
 
     num_requests = 10
     max_tokens = 256
-    generation_cfg = GenerationConfig(max_tokens=max_tokens, n=3)
+    generation_cfg = GenerationConfig(max_tokens=max_tokens, n=7)
 
     output_texts: List[List[str]] = [
         ["" for _ in range(generation_cfg.n)] for _ in range(num_requests)
     ]
 
     async def generate_task(
-        async_engine: AsyncThreadedEngine,
+        async_engine: AsyncEngine,
         prompt: str,
         generation_cfg: GenerationConfig,
         request_id: str,
     ):
         print(f"generate task for request {request_id}")
         rid = int(request_id)
-        async for delta_outputs in async_engine.generate(
+        async for delta_outputs in async_engine._generate(
             prompt, generation_cfg, request_id=request_id
         ):
             assert len(delta_outputs) == generation_cfg.n
@@ -76,5 +76,107 @@ async def test_engine_generate():
     del async_engine
 
 
+async def test_chat_completion():
+    # Initialize model loading info and KV cache config
+    model = ModelInfo(
+        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
+        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
+    )
+    kv_cache_config = KVCacheConfig(page_size=16, max_total_sequence_length=4096)
+    # Create engine
+    async_engine = AsyncEngine(model, kv_cache_config)
+
+    num_requests = 2
+    max_tokens = 32
+    n = 1
+    output_texts: List[List[str]] = [["" for _ in range(n)] for _ in range(num_requests)]
+
+    async def generate_task(prompt: str, request_id: str):
+        print(f"generate chat completion task for request {request_id}")
+        rid = int(request_id)
+        async for response in async_engine.chat_completion(
+            messages=[{"role": "user", "content": prompt}],
+            model=model.model,
+            max_tokens=max_tokens,
+            n=n,
+            request_id=request_id,
+        ):
+            for choice in response.choices:
+                assert choice.delta.role == "assistant"
+                output_texts[rid][choice.index] += choice.delta.content
+
+    tasks = [
+        asyncio.create_task(generate_task(prompts[i], request_id=str(i)))
+        for i in range(num_requests)
+    ]
+
+    await asyncio.gather(*tasks)
+
+    # Print output.
+    print("Chat completion all finished")
+    for req_id, outputs in enumerate(output_texts):
+        print(f"Prompt {req_id}: {prompts[req_id]}")
+        if len(outputs) == 1:
+            print(f"Output {req_id}:{outputs[0]}\n")
+        else:
+            for i, output in enumerate(outputs):
+                print(f"Output {req_id}({i}):{output}\n")
+
+    async_engine.terminate()
+    del async_engine
+
+
+async def test_completion():
+    # Initialize model loading info and KV cache config
+    model = ModelInfo(
+        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
+        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
+    )
+    kv_cache_config = KVCacheConfig(page_size=16, max_total_sequence_length=4096)
+    # Create engine
+    async_engine = AsyncEngine(model, kv_cache_config)
+
+    num_requests = 2
+    max_tokens = 128
+    n = 1
+    output_texts: List[List[str]] = [["" for _ in range(n)] for _ in range(num_requests)]
+
+    async def generate_task(prompt: str, request_id: str):
+        print(f"generate completion task for request {request_id}")
+        rid = int(request_id)
+        async for response in async_engine.completion(
+            prompt=prompt,
+            model=model.model,
+            max_tokens=max_tokens,
+            n=n,
+            ignore_eos=True,
+            request_id=request_id,
+        ):
+            for choice in response.choices:
+                output_texts[rid][choice.index] += choice.text
+
+    tasks = [
+        asyncio.create_task(generate_task(prompts[i], request_id=str(i)))
+        for i in range(num_requests)
+    ]
+
+    await asyncio.gather(*tasks)
+
+    # Print output.
+    print("Chat completion all finished")
+    for req_id, outputs in enumerate(output_texts):
+        print(f"Prompt {req_id}: {prompts[req_id]}")
+        if len(outputs) == 1:
+            print(f"Output {req_id}:{outputs[0]}\n")
+        else:
+            for i, output in enumerate(outputs):
+                print(f"Output {req_id}({i}):{output}\n")
+
+    async_engine.terminate()
+    del async_engine
+
+
 if __name__ == "__main__":
     asyncio.run(test_engine_generate())
+    asyncio.run(test_chat_completion())
+    asyncio.run(test_completion())
