+++ tests/python/serve/test_serve_engine.py
@@ -1,18 +1,9 @@
 # pylint: disable=chained-comparison,line-too-long,missing-docstring,
 # pylint: disable=too-many-arguments,too-many-locals,unused-argument,unused-variable
-from typing import Callable, List, Optional
+from typing import List
 
-import numpy as np
-
-from mlc_llm.serve import (
-    Engine,
-    GenerationConfig,
-    KVCacheConfig,
-    Request,
-    RequestStreamOutput,
-    data,
-)
-from mlc_llm.serve.engine import ModelInfo
+from mlc_llm.serve import Engine, GenerationConfig, KVCacheConfig
+from mlc_llm.serve.engine_base import ModelInfo
 
 prompts = [
     "What is the meaning of life?",
@@ -28,344 +19,6 @@ prompts = [
 ]
 
 
-def create_requests(
-    num_requests: int,
-    stop_token_id: Optional[int] = None,
-    temperature: float = 0.8,
-    repetition_penalty: float = 1.0,
-    max_tokens_low: int = 256,
-    max_tokens_high: int = 257,
-) -> List[Request]:
-    assert num_requests >= 0 and num_requests <= len(prompts)
-
-    stop_token_ids = [stop_token_id] if stop_token_id is not None else []
-    requests = []
-    for req_id, prompt in zip(range(num_requests), prompts):
-        max_tokens = np.random.randint(max_tokens_low, max_tokens_high)
-        requests.append(
-            Request(
-                request_id=str(req_id),
-                inputs=data.TextData(prompt),
-                generation_config=GenerationConfig(
-                    temperature=temperature,
-                    repetition_penalty=repetition_penalty,
-                    max_tokens=max_tokens,
-                    stop_token_ids=stop_token_ids,
-                ),
-            )
-        )
-    return requests
-
-
-def test_engine_basic():
-    """Test engine **without continuous batching**.
-
-    - Add all requests to the engine altogether in the beginning.
-    - All requests have the same max_tokens. This means all requests
-    will end together.
-    - Engine keeps running `step` for estimated number of steps (number of
-    requests + max_tokens - 1). Then check the output of each request.
-    """
-
-    # Initialize model loading info and KV cache config
-    model = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
-    )
-    kv_cache_config = KVCacheConfig(page_size=16)
-
-    # Hyperparameters for tests (you can try different combinations).
-    num_requests = 10  # [4, 8, 10]
-    temperature = 0.9  # [0, 0.8, 0.9, 1.0, 1.1]
-    repetition_penalty = 1.0  # [1.0, 1.01]
-    max_tokens: int = 256  # [32, 128, 256]
-    np.random.seed(0)
-
-    # Output list
-    outputs = [[] for _ in range(num_requests)]
-
-    # Define the callback function for request generation results
-    def fcallback(delta_outputs: List[RequestStreamOutput]):
-        for delta_output in delta_outputs:
-            request_id, stream_outputs = delta_output.unpack()
-            assert len(stream_outputs) == 1
-            outputs[int(request_id)] += stream_outputs[0].delta_token_ids
-
-    # Create engine
-    engine = Engine(model, kv_cache_config, request_stream_callback=fcallback)
-
-    # Create requests
-    requests = create_requests(
-        num_requests,
-        temperature=temperature,
-        repetition_penalty=repetition_penalty,
-        max_tokens_low=max_tokens,
-        max_tokens_high=max_tokens + 1,
-    )
-
-    # Add all requests to engine
-    for request in requests:
-        engine.add_request(request)
-
-    num_steps = num_requests + max_tokens - 1
-    # Run steps
-    for step in range(num_steps):
-        engine.step()
-
-    for req_id, output in enumerate(outputs):
-        print(f"Prompt {req_id}: {requests[req_id].inputs[0]}")
-        print(f"Output {req_id}:{engine.tokenizer.decode(output)}\n")
-
-
-def test_engine_continuous_batching_1():
-    """Test engine **with continuous batching**.
-
-    - Add all requests to the engine altogether in the beginning.
-    - All requests have a random maximum generation length. So each
-    request keeps generating until reaching the maximum length.
-    - Engine keeps running `step` for estimated number of steps (number of
-    requests + the maximum max_tokens - 1). Then check the output
-    of each request.
-    """
-
-    # Initialize model loading info and KV cache config
-    model = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
-    )
-    kv_cache_config = KVCacheConfig(page_size=16)
-
-    # Hyperparameters for tests (you can try different combinations)
-    num_requests = 10  # [4, 8, 10]
-    temperature = 0.9  # [0.8, 0.9, 1.0, 1.1]
-    repetition_penalty = 1.00  # [1.0, 1.01]
-    max_tokens_low = 128
-    max_tokens_high = 384
-    np.random.seed(0)
-
-    # Output list
-    outputs = [[] for _ in range(num_requests)]
-    finish_time = [None] * num_requests
-
-    # Define the callback class for request generation results
-    class CallbackTimer:
-        timer: int = -1
-
-        def callback_getter(self) -> Callable[[List[RequestStreamOutput]], None]:
-            def fcallback(delta_outputs: List[RequestStreamOutput]):
-                for delta_output in delta_outputs:
-                    request_id, stream_outputs = delta_output.unpack()
-                    assert len(stream_outputs) == 1
-                    if stream_outputs[0].finish_reason is not None:
-                        print(f"Request {request_id} finished at step {self.timer}.")
-                    outputs[int(request_id)] += stream_outputs[0].delta_token_ids
-                    finish_time[int(request_id)] = self.timer
-
-            return fcallback
-
-        def step(self) -> None:
-            self.timer += 1
-
-    # Create engine
-    timer = CallbackTimer()
-    engine = Engine(model, kv_cache_config, request_stream_callback=timer.callback_getter())
-
-    # Create requests
-    requests = create_requests(
-        num_requests,
-        temperature=temperature,
-        repetition_penalty=repetition_penalty,
-        max_tokens_low=max_tokens_low,
-        max_tokens_high=max_tokens_high,
-    )
-
-    # Add all requests to engine
-    for request in requests:
-        engine.add_request(request)
-
-    num_steps = num_requests + max(request.generation_config.max_tokens for request in requests) - 1
-    # Run steps
-    for step in range(num_steps):
-        timer.step()
-        assert timer.timer == step
-        engine.step()
-
-    for req_id, (request, output, fin_time) in enumerate(zip(requests, outputs, finish_time)):
-        print(f"Prompt {req_id}: {request.inputs[0]}")
-        print(f"Output {req_id}:{engine.tokenizer.decode(output)}\n")
-        assert fin_time == request.generation_config.max_tokens - 1
-
-
-def test_engine_continuous_batching_2():
-    """Test engine **with continuous batching**.
-
-    - Add all requests to the engine altogether in the beginning.
-    - All requests have the stop token. So each request keeps generating
-    until having the stop token or reaching the maximum length.
-    - Engine keeps running `step` for estimated number of steps (number of
-    requests + the maximum max_tokens - 1). Then check the output
-    of each request.
-    """
-
-    # Initialize model loading info and KV cache config
-    model = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
-    )
-    kv_cache_config = KVCacheConfig(page_size=16)
-
-    # Hyperparameters for tests (you can try different combinations)
-    num_requests = 10  # [4, 8, 10]
-    temperature = 0.9  # [0.8, 0.9, 1.0, 1.1]
-    repetition_penalty = 1.00  # [1.0, 1.01]
-    stop_token_id = 2
-    max_tokens = 512
-    np.random.seed(0)
-
-    # Output list
-    outputs = [[] for _ in range(num_requests)]
-    finish_time = [None] * num_requests
-
-    # Define the callback class for request generation results
-    class CallbackTimer:
-        timer: int = -1
-
-        def callback_getter(self) -> Callable[[List[RequestStreamOutput]], None]:
-            def fcallback(delta_outputs: List[RequestStreamOutput]):
-                for delta_output in delta_outputs:
-                    request_id, stream_outputs = delta_output.unpack()
-                    assert len(stream_outputs) == 1
-                    if stream_outputs[0].finish_reason is not None:
-                        print(f"Request {request_id} finished at step {self.timer}.")
-                    outputs[int(request_id)] += stream_outputs[0].delta_token_ids
-                    finish_time[int(request_id)] = self.timer
-
-            return fcallback
-
-        def step(self) -> None:
-            self.timer += 1
-
-    # Create engine
-    timer = CallbackTimer()
-    engine = Engine(model, kv_cache_config, request_stream_callback=timer.callback_getter())
-
-    # Create requests
-    requests = create_requests(
-        num_requests,
-        stop_token_id=stop_token_id,
-        temperature=temperature,
-        repetition_penalty=repetition_penalty,
-        max_tokens_low=max_tokens,
-        max_tokens_high=max_tokens + 1,
-    )
-
-    # Add all requests to engine
-    for request in requests:
-        engine.add_request(request)
-
-    num_steps = num_requests + max_tokens - 1
-    # Run steps
-    for step in range(num_steps):
-        timer.step()
-        assert timer.timer == step
-        engine.step()
-
-    for req_id, (request, output, fin_time) in enumerate(zip(requests, outputs, finish_time)):
-        print(f"Prompt {req_id}: {request.inputs[0]}")
-        if fin_time < num_requests + max_tokens - 2:
-            print(f"Request {req_id} ends early on the stop token")
-        print(f"Output {req_id}:{engine.tokenizer.decode(output)}\n")
-
-
-def test_engine_continuous_batching_3():
-    """Test engine **with continuous batching**.
-
-    - Add requests randomly between time [0, 200).
-    - All requests have a random maximum generation length. So each
-    request keeps generating until reaching the maximum length.
-    - Engine keeps running `step` until all requests finish.
-    Then check the output of each request.
-    """
-
-    # Initialize model loading info and KV cache config
-    model = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
-    )
-    kv_cache_config = KVCacheConfig(page_size=16)
-
-    # Hyperparameters for tests (you can try different combinations)
-    num_requests = 10  # [4, 8, 10]
-    temperature = 0.9  # [0.8, 0.9, 1.0, 1.1]
-    repetition_penalty = 1.00  # [1.0, 1.01]
-    stop_token_id = 2
-    max_tokens_low = 64
-    max_tokens_high = 192
-    np.random.seed(0)
-
-    # Output list
-    outputs = [[] for _ in range(num_requests)]
-    finish_time = [None] * num_requests
-
-    # Define the callback class for request generation results
-    class CallbackTimer:
-        timer: int = -1
-        finished_requests: int = 0
-
-        def callback_getter(self) -> Callable[[List[RequestStreamOutput]], None]:
-            def fcallback(delta_outputs: List[RequestStreamOutput]):
-                for delta_output in delta_outputs:
-                    request_id, stream_outputs = delta_output.unpack()
-                    assert len(stream_outputs) == 1
-                    if stream_outputs[0].finish_reason is not None:
-                        print(f"Request {request_id} finished at step {self.timer}.")
-                        self.finished_requests += 1
-                    outputs[int(request_id)] += stream_outputs[0].delta_token_ids
-                    finish_time[int(request_id)] = self.timer
-
-            return fcallback
-
-        def step(self) -> None:
-            self.timer += 1
-
-        def all_finished(self) -> bool:
-            return self.finished_requests == num_requests
-
-    # Create engine
-    timer = CallbackTimer()
-    engine = Engine(model, kv_cache_config, request_stream_callback=timer.callback_getter())
-
-    # Create requests
-    requests = create_requests(
-        num_requests,
-        stop_token_id=stop_token_id,
-        temperature=temperature,
-        repetition_penalty=repetition_penalty,
-        max_tokens_low=max_tokens_low,
-        max_tokens_high=max_tokens_high,
-    )
-
-    # Assign the time to add requests to engine
-    request_add_time = [np.random.randint(0, 200) for _ in range(num_requests)]
-
-    # Run steps
-    while not timer.all_finished():
-        timer.step()
-
-        # Add requests to engine
-        for req_id, add_time in enumerate(request_add_time):
-            if add_time == timer.timer:
-                print(f"add request {req_id} at step {timer.timer}")
-                engine.add_request(requests[req_id])
-
-        engine.step()
-
-    for req_id, (request, output, fin_time) in enumerate(zip(requests, outputs, finish_time)):
-        print(f"Prompt {req_id}: {request.inputs[0]}")
-        print(f"Finish time: {fin_time}")
-        print(f"Output {req_id}:{engine.tokenizer.decode(output)}\n")
-
-
 def test_engine_generate():
     # Initialize model loading info and KV cache config
     model = ModelInfo(
@@ -378,11 +31,20 @@ def test_engine_generate():
 
     num_requests = 10
     max_tokens = 256
+    generation_cfg = GenerationConfig(max_tokens=max_tokens, n=7)
 
-    # Generate output.
-    output_texts, _ = engine.generate(
-        prompts[:num_requests], GenerationConfig(max_tokens=max_tokens)
-    )
+    output_texts: List[List[str]] = [
+        ["" for _ in range(generation_cfg.n)] for _ in range(num_requests)
+    ]
+    for rid in range(num_requests):
+        print(f"generating for request {rid}")
+        for delta_outputs in engine._generate(prompts[rid], generation_cfg, request_id=str(rid)):
+            assert len(delta_outputs) == generation_cfg.n
+            for i, delta_output in enumerate(delta_outputs):
+                output_texts[rid][i] += delta_output.delta_text
+
+    # Print output.
+    print("All finished")
     for req_id, outputs in enumerate(output_texts):
         print(f"Prompt {req_id}: {prompts[req_id]}")
         if len(outputs) == 1:
@@ -391,10 +53,95 @@ def test_engine_generate():
             for i, output in enumerate(outputs):
                 print(f"Output {req_id}({i}):{output}\n")
 
+    engine.terminate()
+    del engine
+
+
+def test_chat_completion():
+    # Initialize model loading info and KV cache config
+    model = ModelInfo(
+        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
+        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
+    )
+    kv_cache_config = KVCacheConfig(page_size=16, max_total_sequence_length=4096)
+    # Create engine
+    engine = Engine(model, kv_cache_config)
+
+    num_requests = 2
+    max_tokens = 64
+    n = 2
+    output_texts: List[List[str]] = [["" for _ in range(n)] for _ in range(num_requests)]
+
+    for rid in range(num_requests):
+        print(f"chat completion for request {rid}")
+        for response in engine.chat_completion(
+            messages=[{"role": "user", "content": prompts[rid]}],
+            model=model.model,
+            max_tokens=max_tokens,
+            n=n,
+            request_id=str(rid),
+        ):
+            for choice in response.choices:
+                assert choice.delta.role == "assistant"
+                output_texts[rid][choice.index] += choice.delta.content
+
+    # Print output.
+    print("Chat completion all finished")
+    for req_id, outputs in enumerate(output_texts):
+        print(f"Prompt {req_id}: {prompts[req_id]}")
+        if len(outputs) == 1:
+            print(f"Output {req_id}:{outputs[0]}\n")
+        else:
+            for i, output in enumerate(outputs):
+                print(f"Output {req_id}({i}):{output}\n")
+
+    engine.terminate()
+    del engine
+
+
+def test_completion():
+    # Initialize model loading info and KV cache config
+    model = ModelInfo(
+        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
+        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
+    )
+    kv_cache_config = KVCacheConfig(page_size=16, max_total_sequence_length=4096)
+    # Create engine
+    engine = Engine(model, kv_cache_config)
+
+    num_requests = 2
+    max_tokens = 128
+    n = 1
+    output_texts: List[List[str]] = [["" for _ in range(n)] for _ in range(num_requests)]
+
+    for rid in range(num_requests):
+        print(f"completion for request {rid}")
+        for response in engine.completion(
+            prompt=prompts[rid],
+            model=model.model,
+            max_tokens=max_tokens,
+            n=n,
+            ignore_eos=True,
+            request_id=str(rid),
+        ):
+            for choice in response.choices:
+                output_texts[rid][choice.index] += choice.text
+
+    # Print output.
+    print("Chat completion all finished")
+    for req_id, outputs in enumerate(output_texts):
+        print(f"Prompt {req_id}: {prompts[req_id]}")
+        if len(outputs) == 1:
+            print(f"Output {req_id}:{outputs[0]}\n")
+        else:
+            for i, output in enumerate(outputs):
+                print(f"Output {req_id}({i}):{output}\n")
+
+    engine.terminate()
+    del engine
+
 
 if __name__ == "__main__":
-    test_engine_basic()
-    test_engine_continuous_batching_1()
-    test_engine_continuous_batching_2()
-    test_engine_continuous_batching_3()
     test_engine_generate()
+    test_chat_completion()
+    test_completion()
