+++ tests/python/serve/test_serve_engine_image.py
@@ -1,10 +1,13 @@
-from mlc_llm.serve import Engine, GenerationConfig, KVCacheConfig, data
-from mlc_llm.serve.engine import ModelInfo
-from mlc_llm.serve.entrypoints.entrypoint_utils import get_image_from_url
+import json
+from pathlib import Path
+
+from mlc_llm.serve import GenerationConfig, KVCacheConfig, data
+from mlc_llm.serve.engine_base import ModelInfo
+from mlc_llm.serve.sync_engine import SyncEngine
 
 
-def get_test_image():
-    return get_image_from_url("https://llava-vl.github.io/static/images/view.jpg")
+def get_test_image(config) -> data.ImageData:
+    return data.ImageData.from_url("https://llava-vl.github.io/static/images/view.jpg", config)
 
 
 def test_engine_generate():
@@ -15,19 +18,21 @@ def test_engine_generate():
     )
     kv_cache_config = KVCacheConfig(page_size=16, max_total_sequence_length=4096)
     # Create engine
-    engine = Engine(model, kv_cache_config)
-
+    engine = SyncEngine(model, kv_cache_config)
     max_tokens = 256
+
+    with open(Path(model.model) / "mlc-chat-config.json", "r", encoding="utf-8") as file:
+        model_config = json.load(file)
 
     prompts = [
         [
             data.TextData("USER: "),
-            data.ImageData(get_test_image(), 576),
+            get_test_image(model_config),
             data.TextData("\nWhat does this image represent? ASSISTANT:"),
         ],
         [
             data.TextData("USER: "),
-            data.ImageData(get_test_image(), 576),
+            get_test_image(model_config),
             data.TextData("\nIs there a dog in this image? ASSISTANT:"),
         ],
         [data.TextData("USER: What is the meaning of life? ASSISTANT:")],
