+++ tests/python/serve/test_serve_sync_engine.py
@@ -5,7 +5,6 @@ from typing import Callable, List, Optional
 import numpy as np
 
 from mlc_llm.serve import (
-    EngineMode,
     GenerationConfig,
     KVCacheConfig,
     Request,
@@ -69,19 +68,14 @@ def test_engine_basic():
     """
 
     # Initialize model loading info and KV cache config
-    ssm = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q4f16_1-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q4f16_1-MLC/Llama-2-7b-chat-hf-q4f16_1-MLC-cuda.so",
-    )
     model = ModelInfo(
         "dist/Llama-2-7b-chat-hf-q0f16-MLC",
         model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
     )
     kv_cache_config = KVCacheConfig(page_size=16)
-    engine_mode = EngineMode(enable_speculative=True)
 
     # Hyperparameters for tests (you can try different combinations).
-    num_requests = len(prompts)  # [4, 8, 10]
+    num_requests = 10  # [4, 8, 10]
     temperature = 0.9  # [0, 0.8, 0.9, 1.0, 1.1]
     repetition_penalty = 1.0  # [1.0, 1.01]
     max_tokens: int = 256  # [32, 128, 256]
@@ -98,7 +92,7 @@ def test_engine_basic():
             outputs[int(request_id)] += stream_outputs[0].delta_token_ids
 
     # Create engine
-    engine = SyncEngine([model, ssm], kv_cache_config, engine_mode, fcallback)
+    engine = SyncEngine(model, kv_cache_config, request_stream_callback=fcallback)
 
     # Create requests
     requests = create_requests(
@@ -135,19 +129,14 @@ def test_engine_continuous_batching_1():
     """
 
     # Initialize model loading info and KV cache config
-    ssm = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q4f16_1-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q4f16_1-MLC/Llama-2-7b-chat-hf-q4f16_1-MLC-cuda.so",
-    )
     model = ModelInfo(
         "dist/Llama-2-7b-chat-hf-q0f16-MLC",
         model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
     )
     kv_cache_config = KVCacheConfig(page_size=16)
-    engine_mode = EngineMode(enable_speculative=True)
 
     # Hyperparameters for tests (you can try different combinations)
-    num_requests = len(prompts)  # [4, 8, 10]
+    num_requests = 10  # [4, 8, 10]
     temperature = 0.9  # [0.8, 0.9, 1.0, 1.1]
     repetition_penalty = 1.00  # [1.0, 1.01]
     max_tokens_low = 128
@@ -179,7 +168,7 @@ def test_engine_continuous_batching_1():
 
     # Create engine
     timer = CallbackTimer()
-    engine = SyncEngine([model, ssm], kv_cache_config, engine_mode, timer.callback_getter())
+    engine = SyncEngine(model, kv_cache_config, request_stream_callback=timer.callback_getter())
 
     # Create requests
     requests = create_requests(
@@ -204,73 +193,68 @@ def test_engine_continuous_batching_1():
     for req_id, (request, output, fin_time) in enumerate(zip(requests, outputs, finish_time)):
         print(f"Prompt {req_id}: {request.inputs[0]}")
         print(f"Output {req_id}:{engine.tokenizer.decode(output)}\n")
-        # assert fin_time == request.generation_config.max_tokens - 1
+        assert (
+            fin_time == request.generation_config.max_tokens - 1
+        ), f"finish time = {fin_time}, max tokens = {request.generation_config.max_tokens - 1}"
 
 
-def test_engine_generate():
+def test_engine_continuous_batching_2():
+    """Test engine **with continuous batching**.
+
+    - Add all requests to the engine altogether in the beginning.
+    - All requests have the stop token. So each request keeps generating
+    until having the stop token or reaching the maximum length.
+    - Engine keeps running `step` for estimated number of steps (number of
+    requests + the maximum max_tokens - 1). Then check the output
+    of each request.
+    """
+
     # Initialize model loading info and KV cache config
-    ssm = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q4f16_1-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q4f16_1-MLC/Llama-2-7b-chat-hf-q4f16_1-MLC-cuda.so",
-    )
     model = ModelInfo(
         "dist/Llama-2-7b-chat-hf-q0f16-MLC",
         model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
     )
     kv_cache_config = KVCacheConfig(page_size=16)
-    engine_mode = EngineMode(enable_speculative=True)
-    # Create engine
-    engine = SyncEngine([model, ssm], kv_cache_config, engine_mode)
 
-    num_requests = 10
-    max_tokens = 256
-
-    # Generate output.
-    output_texts, _ = engine.generate(
-        prompts[:num_requests], GenerationConfig(max_tokens=max_tokens, n=3)
-    )
-    for req_id, outputs in enumerate(output_texts):
-        print(f"Prompt {req_id}: {prompts[req_id]}")
-        if len(outputs) == 1:
-            print(f"Output {req_id}:{outputs[0]}\n")
-        else:
-            for i, output in enumerate(outputs):
-                print(f"Output {req_id}({i}):{output}\n")
-
-
-def test_engine_efficiency():
-    """Test engine speculative decoding efficiency."""
-
-    # Initialize model loading info and KV cache config
-    model = ModelInfo(
-        "dist/Llama-2-13b-chat-hf-q4f16_1-MLC",
-        model_lib_path="dist/Llama-2-13b-chat-hf-q4f16_1-MLC/Llama-2-13b-chat-hf-q4f16_1-MLC-cuda.so",
-    )
-    kv_cache_config = KVCacheConfig(page_size=16)
-
-    # Hyperparameters for tests (you can try different combinations).
-    num_requests = 1  # [4, 8, 10]
-    temperature = 0.9  # [0, 0.8, 0.9, 1.0, 1.1]
-    repetition_penalty = 1.0  # [1.0, 1.01]
-    max_tokens: int = 512
+    # Hyperparameters for tests (you can try different combinations)
+    num_requests = 10  # [4, 8, 10]
+    temperature = 0.9  # [0.8, 0.9, 1.0, 1.1]
+    repetition_penalty = 1.00  # [1.0, 1.01]
+    stop_token_id = 2
+    max_tokens = 512
     np.random.seed(0)
 
     # Output list
     outputs = [[] for _ in range(num_requests)]
+    finish_time = [None] * num_requests
 
-    # Define the callback function for request generation results
-    def fcallback(delta_outputs: List[RequestStreamOutput]):
-        for delta_output in delta_outputs:
-            request_id, stream_outputs = delta_output.unpack()
-            assert len(stream_outputs) == 1
-            outputs[int(request_id)] += stream_outputs[0].delta_token_ids
+    # Define the callback class for request generation results
+    class CallbackTimer:
+        timer: int = -1
+
+        def callback_getter(self) -> Callable[[List[RequestStreamOutput]], None]:
+            def fcallback(delta_outputs: List[RequestStreamOutput]):
+                for delta_output in delta_outputs:
+                    request_id, stream_outputs = delta_output.unpack()
+                    assert len(stream_outputs) == 1
+                    if stream_outputs[0].finish_reason is not None:
+                        print(f"Request {request_id} finished at step {self.timer}.")
+                    outputs[int(request_id)] += stream_outputs[0].delta_token_ids
+                    finish_time[int(request_id)] = self.timer
+
+            return fcallback
+
+        def step(self) -> None:
+            self.timer += 1
 
     # Create engine
-    engine = SyncEngine(model, kv_cache_config, request_stream_callback=fcallback)
+    timer = CallbackTimer()
+    engine = SyncEngine(model, kv_cache_config, request_stream_callback=timer.callback_getter())
 
     # Create requests
     requests = create_requests(
         num_requests,
+        stop_token_id=stop_token_id,
         temperature=temperature,
         repetition_penalty=repetition_penalty,
         max_tokens_low=max_tokens,
@@ -284,97 +268,135 @@ def test_engine_efficiency():
     num_steps = num_requests + max_tokens - 1
     # Run steps
     for step in range(num_steps):
+        timer.step()
+        assert timer.timer == step
         engine.step()
 
-    for eg, name in zip([engine], ["Normal Deconding"]):
-        stats = eg.stats()
-        print("engine name:", name)
-        if name == "Speculative Decoding":
-            print("total draft tokens:", stats["total_draft_tokens"])
-            print("total accepted tokens:", stats["total_accepted_tokens"])
-            print(
-                "Accept rate:",
-                stats["total_accepted_tokens"] / (1e-10 + stats["total_draft_tokens"]),
-            )
-        print("engine total decode time:", stats["engine_total_decode_time"])
-        print()
+    for req_id, (request, output, fin_time) in enumerate(zip(requests, outputs, finish_time)):
+        print(f"Prompt {req_id}: {request.inputs[0]}")
+        if fin_time < num_requests + max_tokens - 2:
+            print(f"Request {req_id} ends early on the stop token")
+        print(f"Output {req_id}:{engine.tokenizer.decode(output)}\n")
 
 
-def test_engine_spec_efficiency():
-    """Test engine speculative decoding efficiency."""
+def test_engine_continuous_batching_3():
+    """Test engine **with continuous batching**.
+
+    - Add requests randomly between time [0, 200).
+    - All requests have a random maximum generation length. So each
+    request keeps generating until reaching the maximum length.
+    - Engine keeps running `step` until all requests finish.
+    Then check the output of each request.
+    """
 
     # Initialize model loading info and KV cache config
-    ssm = ModelInfo(
-        "dist/Llama-2-7b-chat-hf-q4f16_1-MLC",
-        model_lib_path="dist/Llama-2-7b-chat-hf-q4f16_1-MLC/Llama-2-7b-chat-hf-q4f16_1-MLC-cuda.so",
-    )
-    # If Flashinfer allows head_dim < 128, we can test this model
-    # ssm = ModelInfo(
-    #     "dist/TinyLlama-1.1B-Chat-v1.0-q0f16-MLC",
-    #     model_lib_path="dist/TinyLlama-1.1B-Chat-v1.0-q0f16-MLC/TinyLlama-1.1B-Chat-v1.0-q0f16-MLC-cuda.so",
-    # )
     model = ModelInfo(
-        "dist/Llama-2-13b-chat-hf-q4f16_1-MLC",
-        model_lib_path="dist/Llama-2-13b-chat-hf-q4f16_1-MLC/Llama-2-13b-chat-hf-q4f16_1-MLC-cuda.so",
+        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
+        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
     )
     kv_cache_config = KVCacheConfig(page_size=16)
-    engine_mode = EngineMode(enable_speculative=True, spec_draft_length=6)
 
-    # Hyperparameters for tests (you can try different combinations).
-    num_requests = 1  # [4, 8, 10]
-    temperature = 0.9  # [0, 0.8, 0.9, 1.0, 1.1]
-    repetition_penalty = 1.0  # [1.0, 1.01]
-    max_tokens: int = 512
+    # Hyperparameters for tests (you can try different combinations)
+    num_requests = 10  # [4, 8, 10]
+    temperature = 0.9  # [0.8, 0.9, 1.0, 1.1]
+    repetition_penalty = 1.00  # [1.0, 1.01]
+    stop_token_id = 2
+    max_tokens_low = 64
+    max_tokens_high = 192
     np.random.seed(0)
 
     # Output list
     outputs = [[] for _ in range(num_requests)]
+    finish_time = [None] * num_requests
 
-    # Define the callback function for request generation results
-    def fcallback(delta_outputs: List[RequestStreamOutput]):
-        for delta_output in delta_outputs:
-            request_id, stream_outputs = delta_output.unpack()
-            assert len(stream_outputs) == 1
-            outputs[int(request_id)] += stream_outputs[0].delta_token_ids
+    # Define the callback class for request generation results
+    class CallbackTimer:
+        timer: int = -1
+        finished_requests: int = 0
+
+        def callback_getter(self) -> Callable[[List[RequestStreamOutput]], None]:
+            def fcallback(delta_outputs: List[RequestStreamOutput]):
+                for delta_output in delta_outputs:
+                    request_id, stream_outputs = delta_output.unpack()
+                    assert len(stream_outputs) == 1
+                    if stream_outputs[0].finish_reason is not None:
+                        print(f"Request {request_id} finished at step {self.timer}.")
+                        self.finished_requests += 1
+                    outputs[int(request_id)] += stream_outputs[0].delta_token_ids
+                    finish_time[int(request_id)] = self.timer
+
+            return fcallback
+
+        def step(self) -> None:
+            self.timer += 1
+
+        def all_finished(self) -> bool:
+            return self.finished_requests == num_requests
 
     # Create engine
-    spec_engine = SyncEngine([model, ssm], kv_cache_config, engine_mode, fcallback)
+    timer = CallbackTimer()
+    engine = SyncEngine(model, kv_cache_config, request_stream_callback=timer.callback_getter())
 
     # Create requests
     requests = create_requests(
         num_requests,
+        stop_token_id=stop_token_id,
         temperature=temperature,
         repetition_penalty=repetition_penalty,
-        max_tokens_low=max_tokens,
-        max_tokens_high=max_tokens + 1,
+        max_tokens_low=max_tokens_low,
+        max_tokens_high=max_tokens_high,
     )
 
-    # Add all requests to engine
-    for request in requests:
-        spec_engine.add_request(request)
+    # Assign the time to add requests to engine
+    request_add_time = [np.random.randint(0, 200) for _ in range(num_requests)]
 
-    num_steps = num_requests + max_tokens - 1
     # Run steps
-    for step in range(num_steps):
-        spec_engine.step()
+    while not timer.all_finished():
+        timer.step()
 
-    for eg, name in zip([spec_engine], ["Speculative Decoding"]):
-        stats = eg.stats()
-        print("engine name:", name)
-        if name == "Speculative Decoding":
-            print("total draft tokens:", stats["total_draft_tokens"])
-            print("total accepted tokens:", stats["total_accepted_tokens"])
-            print(
-                "Accept rate:",
-                stats["total_accepted_tokens"] / (1e-10 + stats["total_draft_tokens"]),
-            )
-        print("engine total decode time:", stats["engine_total_decode_time"])
-        print()
+        # Add requests to engine
+        for req_id, add_time in enumerate(request_add_time):
+            if add_time == timer.timer:
+                print(f"add request {req_id} at step {timer.timer}")
+                engine.add_request(requests[req_id])
+
+        engine.step()
+
+    for req_id, (request, output, fin_time) in enumerate(zip(requests, outputs, finish_time)):
+        print(f"Prompt {req_id}: {request.inputs[0]}")
+        print(f"Finish time: {fin_time}")
+        print(f"Output {req_id}:{engine.tokenizer.decode(output)}\n")
+
+
+def test_engine_generate():
+    # Initialize model loading info and KV cache config
+    model = ModelInfo(
+        "dist/Llama-2-7b-chat-hf-q0f16-MLC",
+        model_lib_path="dist/Llama-2-7b-chat-hf-q0f16-MLC/Llama-2-7b-chat-hf-q0f16-MLC-cuda.so",
+    )
+    kv_cache_config = KVCacheConfig(page_size=16, max_total_sequence_length=4096)
+    # Create engine
+    engine = SyncEngine(model, kv_cache_config)
+
+    num_requests = 10
+    max_tokens = 256
+
+    # Generate output.
+    output_texts, _ = engine.generate(
+        prompts[:num_requests], GenerationConfig(max_tokens=max_tokens, n=7)
+    )
+    for req_id, outputs in enumerate(output_texts):
+        print(f"Prompt {req_id}: {prompts[req_id]}")
+        if len(outputs) == 1:
+            print(f"Output {req_id}:{outputs[0]}\n")
+        else:
+            for i, output in enumerate(outputs):
+                print(f"Output {req_id}({i}):{output}\n")
 
 
 if __name__ == "__main__":
     test_engine_basic()
     test_engine_continuous_batching_1()
+    test_engine_continuous_batching_2()
+    test_engine_continuous_batching_3()
     test_engine_generate()
-    test_engine_efficiency()
-    test_engine_spec_efficiency()
